Fixes to Batterysense and Menu

+++ CO2_Gadget_Battery.h
@@ -9,8 +9,8 @@
 // clang-format on
 
 	// Battery info (percent charge).
-	const uint16_t voltageDividerR1Ohms = 100000; // Resistance of "R1" for voltage divider.
-	const uint16_t voltageDividerR2Ohms = 100000; // Resistance of "R2" for voltage divider.	
+	const uint32_t voltageDividerR1Ohms = 100000; // Resistance of "R1" for voltage divider.
+	const uint32_t voltageDividerR2Ohms = 100000; // Resistance of "R2" for voltage divider.	
 
   // Voltage divider ratio for battery sense must be (R1 + R2) / R2 - see https://github.com/rlogiacco/BatterySense.
   const float voltageDividerRatio = (voltageDividerR1Ohms + voltageDividerR2Ohms) / voltageDividerR2Ohms;
